Rename all EDD\Database\Tables files. #6331
