+++ includes/database/tables/class-logs-file-downloads.php
